requirements.txt
selenium
webdriver-manager
streamlit
pandas
plotly
openpyxl
python-dotenv
faker